+++ example/lib/main.dart
@@ -30,58 +30,77 @@ class _MyAppState extends State<MyApp> {
           children: [
             DownloadMediaBuilder(
               url: 'https://sample-videos.com/video123/mp4/720/big_buck_bunny_720p_5mb.mp4',
-              onInit: (controller) => this.controller = controller,
-              encryptionPassword: "this is another password",
+              autoDownload: false,
+              onInitialize: (controller) => this.controller = controller,
+              onInitial: (snapshot) {
+                return ElevatedButton(
+                  onPressed: controller.getFile,
+                  child: const Text('Load file'),
+                );
+              },
               onSuccess: (snapshot) {
                 return BetterPlayer.file(snapshot.filePath!);
               },
               onLoading: (snapshot) {
-                return Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: [
-                    LinearProgressIndicator(
-                      value: snapshot.progress,
-                    ),
-                    const SizedBox(height: 20),
-                    ElevatedButton(
-                      onPressed: controller.cancelDownload,
-                      child: const Text('Cancel Download'),
-                    ),
-                  ],
-                );
-              },
-              onCancel: (snapshot) {
-                return ElevatedButton(
-                  onPressed: controller.retry,
-                  child: const Text('Retry'),
-                );
-              },
-              onDecrypting: (snapshot) {
-                return const Center(
-                  child: Text('File is under decryption...'),
-                );
-              },
-              onEncrypting: (snapshot) {
-                return const Center(
-                  child: Text('File is under encryption...'),
-                );
-              },
-              onError: (snapshot) {
-                return Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: [
-                    const Center(child: Text('Error Occurred!')),
-                    const SizedBox(height: 10),
-                    ElevatedButton(
-                      onPressed: controller.retry,
-                      child: const Text('Retry'),
-                    ),
-                  ],
+                return LinearProgressIndicator(
+                  value: snapshot.progress,
                 );
               },
             ),
+            // DownloadMediaBuilder(
+            //   url: 'https://sample-videos.com/video123/mp4/720/big_buck_bunny_720p_5mb.mp4',
+            //   onInitialize: (controller) => this.controller = controller,
+            //   encryptionPassword: "this is another password",
+            //   onSuccess: (snapshot) {
+            //     return BetterPlayer.file(snapshot.filePath!);
+            //   },
+            //   onLoading: (snapshot) {
+            //     return Column(
+            //       mainAxisAlignment: MainAxisAlignment.center,
+            //       crossAxisAlignment: CrossAxisAlignment.stretch,
+            //       children: [
+            //         LinearProgressIndicator(
+            //           value: snapshot.progress,
+            //         ),
+            //         const SizedBox(height: 20),
+            //         ElevatedButton(
+            //           onPressed: controller.cancelDownload,
+            //           child: const Text('Cancel Download'),
+            //         ),
+            //       ],
+            //     );
+            //   },
+            //   onCancel: (snapshot) {
+            //     return ElevatedButton(
+            //       onPressed: controller.retry,
+            //       child: const Text('Retry'),
+            //     );
+            //   },
+            //   onDecrypting: (snapshot) {
+            //     return const Center(
+            //       child: Text('File is under decryption...'),
+            //     );
+            //   },
+            //   onEncrypting: (snapshot) {
+            //     return const Center(
+            //       child: Text('File is under encryption...'),
+            //     );
+            //   },
+            //   onError: (snapshot) {
+            //     return Column(
+            //       mainAxisAlignment: MainAxisAlignment.center,
+            //       crossAxisAlignment: CrossAxisAlignment.stretch,
+            //       children: [
+            //         const Center(child: Text('Error Occurred!')),
+            //         const SizedBox(height: 10),
+            //         ElevatedButton(
+            //           onPressed: controller.retry,
+            //           child: const Text('Retry'),
+            //         ),
+            //       ],
+            //     );
+            //   },
+            // ),
           ],
         ),
       ),