+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.1
+* Rename onInit callback to onInitialize.
+* Add initial state to DownloadMediaStatus.
+
 ## 0.2.0
 * Encrypt and decrypt downloaded files with AES.
 * Encrypt Single File.